Can now edit saved weapon effects.
