+++ src/HiddenPages/WeaponEffectEditor/weapon-effect-tag.css
