+++ src/index.tsx
@@ -9,9 +9,9 @@ import { DialogueTreeMaker } from './RPGTools/DialogueTreeMaker/DialogueTreeMake
 import { CharacterMaker } from './RPGTools/CharacterMaker/CharacterMaker';
 import { NotFoundPage } from './NotFoundPage';
 import { TerminalPage } from './Terminal/TerminalPage';
-import { WeaponEffectForm } from './HiddenPages/WeaponMaker/WeaponEffectForm';
 import { WeaponMaker } from './DNDTools/WeaponMaker/WeaponMaker';
 import { WeaponEffectList } from './HiddenPages/WeaponEffectList/WeaponEffectList';
+import { WeaponEffectEditor } from './HiddenPages/WeaponEffectEditor/WeaponEffectEditor';
 
 const rootDomNode = document.getElementById('app');
 
@@ -39,7 +39,7 @@ const router = createBrowserRouter([
         ]
     },
     { path: "/terminal", element: <TerminalPage /> },
-    { path: "/unlisted/weapon_effect", element: <WeaponEffectForm /> },
+    { path: "/unlisted/weapon_effect", element: <WeaponEffectEditor /> },
     { path: "/unlisted/weapon_effects", element: <WeaponEffectList /> },
     { path: '*', element: <NotFoundPage /> }
 ]);
